+++ client/src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Gallery from './Gallery.jsx';
+import Gallery from './components/Gallery.jsx';
 
 ReactDOM.render(<Gallery />, document.getElementById('gallery'));
